v1/requirements.txt
# GameGen v1 Requirements
# Generated from niya1 conda environment

# Core dependencies for GameGen
pydantic==2.11.7
pillow==11.2.1
pyyaml==6.0.2

# AI/LLM related packages
openai==1.93.0
zhipuai==2.1.5.20250626
langchain==0.3.26
langchain-aws==0.2.27
langchain-community==0.3.26
langchain-core==0.3.66
langchain-deepseek==0.1.3
langchain-openai==0.3.27
langchain-text-splitters
langchainhub==0.1.21
langgraph==0.3.34
langgraph-checkpoint==2.1.0
langgraph-codeact==0.1.3
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.72
langsmith==0.4.4

# AWS/Boto3 for AWS services
boto3==1.38.46
botocore==1.38.46
s3transfer==0.13.0

# HTTP and networking
aiohttp==3.12.13
aiosignal==1.3.2
async-timeout
aiohappyeyeballs==2.6.1
httpx
httpx-sse==0.4.1
httpcore
requests==2.32.4
requests-toolbelt

# Data processing and utilities
numpy==1.26.4
opencv-python==4.11.0.86
tiktoken==0.9.0
regex==2024.11.6
tqdm
python-dateutil==2.9.0.post0
python-dotenv==1.1.1

# JSON and serialization
dataclasses-json==0.6.7
orjson
ormsgpack==1.10.0
jsonpatch
jsonpointer==2.1

# Type hints and validation
typing-extensions
typing-inspect==0.9.0
typing-inspection
mypy_extensions==1.1.0
annotated-types
pydantic-settings==2.10.1
pydantic_core

# Authentication and security
PyJWT==2.8.0
certifi
cffi
pycparser

# Other utilities
attrs==25.3.0
cachetools==6.1.0
charset-normalizer
distro
exceptiongroup
frozenlist==1.7.0
greenlet
h11
idna
jiter
jmespath==1.0.1
marshmallow==3.26.1
multidict==6.5.1
packaging
propcache==0.3.2
PySocks
six==1.17.0
sniffio
SQLAlchemy
tenacity
types-requests==2.32.4.20250611
urllib3
xxhash==3.5.0
yarl==1.20.1
zstandard 